+++ packages/calcite-design-tokens/support/run.ts
@@ -1,6 +1,6 @@
 // @ts-strict-ignore
-import { run as styleDictionaryRunner } from "./token-transformer/sd-run.js";
 import { config } from "../src/$config.js";
+import { run as styleDictionaryRunner } from "./token-transformer/sd-run.js";
 import { CalciteTokenTransformConfig } from "./types/config.js";
 
 const runConfig: CalciteTokenTransformConfig = config;